+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 
 **koot**
 
+-   **重大改动**
+    -   _Node.js_ 最低版本要求提升到 `8.12.0`
+    -   [SSR & 客户端渲染] 组件 CSS 的 `<style>` 标签上不再有 `id` 属性，以避免和元素冲突
+        -   如果项目中有用到根据 `id` 选择 `<style>` 标签的场景，可改为选择标签属性 `[data-koot-module]`
+    -   [服务器] 服务器端打包代码不再会对 `await` `async` 进行转译 ([#172](https://github.com/cmux/koot/issues/172), [#186](https://github.com/cmux/koot/issues/186))
 -   **新特性**
     -   **新工具函数** `koot/utils/client-get-styles` - 获取当前全局 CSS 和所有组件 CSS。具体用法请参见文档 [全局与工具函数/工具函数/客户端](https://koot.js.org/#/utilities?id=客户端) ([#185](https://github.com/cmux/koot/issues/185))
--   优化
-    -   组件 CSS 的 `<style>` 标签上不再有 `id` 属性，以避免和元素冲突
 
 ## [0.11.15] - 2019-10-15